split commandline tests into two files
