+++ run-tests.sh
@@ -29,8 +29,9 @@
 # SPDX-License-Identifier: BSD-2-Clause
 # =================================================================================================
 
-pipenv run pytest --capture=tee-sys  --junitxml=test-results/cmdline-tests.xml tests/test_cmdline.py
 pipenv run pytest --junitxml=test-results/protocol-tests.xml tests/test_protocol.py
 pipenv run pytest --junitxml=test-results/parser-tests.xml tests/test_parsers.py
+pipenv run pytest --capture=tee-sys  --junitxml=test-results/cmdline-dir-tests.xml tests/test_cmdline_dir.py
+pipenv run pytest --capture=tee-sys  --junitxml=test-results/cmdline-date-tests.xml tests/test_cmdline_date.py
 #pipenv run pytest --junitxml=test-results/code-generator-tests.xml tests/output_formatters/test_code_generator.py
 mypy npt/*.py --junit-xml test-results/npt-typecheck.xml
